requirements_CPU.txt
xlrd==1.2.0
pandas==0.24.2
biopython==1.73
matplotlib==3.1.0
numpy==1.16.4
requests==2.22.0
scikit-learn==0.21.2
scipy==1.2.1
click==7.0
configparser==3.7.4
pyfiglet==0.8.post1
keras==2.2.4
tensorflow==1.13.1


